+++ src/libft/Makefile
@@ -59,9 +59,7 @@ SRCS = ft_memset.c\
 		ft_dlstadd_front.c\
 		ft_dlstdelone.c\
 		ft_dlstclear.c\
-		ft_dlstinsert.c\
-		push.c\
-		pop.c
+		ft_dlstinsert.c
 
 OBJS = ${SRCS:.c=.o}
