+++ src/queue/pop.c
